Move thumbnail to components directory
