+++ src/components/thumbnail/Thumbnail.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import './App.css';
+import '../../App.css';
 
 function Thumbnail(props) {
   return (
